+++ ChangeLog.md
@@ -5,5 +5,6 @@
 - 2020-02-12: New reps, 10 minutes
 - 2020-02-13: No reps :(
 - 2020-02-14: Reps, used foldMap instead of foldr for foldable, 8 minutes
+- 2020-02-21: Reps, 10 min. Missed several days. Added Semigroup and Monoid